requirements.txt
#asyncio == 3.4.3
python-binance == 1.0.16
pandas == 1.5.3